requirements.txt
# Install the required dependencies for the project, ensuring compatibility between libraries.

# Set Python version (Colab already provides Python, so this line may not be necessary)
#Python==3.10  # Ensure Python 3.10 for compatibility

# Install PyTorch and related libraries
torch==2.1.2
torchvision==0.16.2
torchtext==0.16.2
torcheval==0.0.7
torchmetrics==0.11.4

# Install essential scientific computing and visualization libraries
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=9.0.0
tqdm>=4.65.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
